ported to Scala 2.11, all test pass but still there are a lot of warnings
