+++ .gitignore
@@ -1,8 +1,24 @@
-*.class
-lib/
-doc/apidoc/
-target/
-lib_managed/
-src_managed/
-project/boot/
-output/
+target
+project/project
+project/target
+.project
+.classpath
+.settings
+bin/
+.cache
+.idea/
+logs
+tmp
+.history
+dist
+/.idea
+/*.iml
+/out
+/.idea_modules
+/.classpath
+/RUNNING_PID
+/.settings
+/project/*-shim.sbt
+/activator-sbt-atmos-akka-shim.sbt
+/~*
+src/test/learn.sc
