--- INSTALL
@@ -1,90 +1,0 @@
-= INSTALL BioScala =
-
-It is rather easy to install a working repository of BioScala. The JVM is
-needed and sbt (simple build tool).
-
-First download BioScala from github with:
-
-  git clone http://github.com/bioscala/bioscala.git
-
-Install Java. On Debian, for example:
-
-  apt-get install sun-java6-jdk 
-
-Install sbt (Simple build tool) from 
-
-  http://code.google.com/p/simple-build-tool/wiki/Setup. 
-
-Download sbt-launch.jar and run it on the command line, for example
-
-  # go to base bioscala directory
-  cd bioscala
-  # launch sbt
-  java -Xmx512M -jar sbt-launch.jar
-  # inside sbt type
-  > update
-  > compile
-  > test
-
-sbt will automatically download Scala and related jars. Then it will compile
-and run the supplied tests. When running sbt from the command line in the base
-dir:
-
-  > test
-  [info] == test-finish ==
-  [info] Passed: : Total 84, Failed 0, Errors 0, Passed 84, Skipped 0
-  [success] Successful.
-  [info] Total time: 2 s
-
-To check current versions and dependencies, see 
-
-  https://github.com/bioscala/bioscala/blob/master/project/build.properties
-  https://github.com/bioscala/bioscala/blob/master/project/build/bioscala.scala 
-
-To try BioScala from other code, build the class files and add them to the
-class path. A jar file will be provided in the future.
-
-= Dependencies =
-
-You need to have:
-
-* BioJAVA:   Add the biojava.jar file to the class path (automatic by sbt)
-
-* scalatest: Add the scalatest.jar file to the class path (automatic by sbt)
-
-Optional are:
-
-* BioRuby:   Unpack the BioRuby source and point the Ruby load path to the
-             lib folder therein, e.g.:
-
-               export RUBYLIB=~/local/lib/bioruby/lib
-
-* JRuby:     Install jruby and add the big jruby-complete.jar file to the 
-             class path, like jruby-complete-1.5.0.RC1.jar (check the version
-             of your JRuby with the -v switch: jruby -v)
-
-* BioJava3:
-
-  biojava3-core-3.0-SNAPSHOT.jar       
-  biojava3-alignment-3.0-SNAPSHOT.jar  
-  biojava3-genome-3.0-SNAPSHOT.jar (optional)
-  biojava3-phylo-3.0-SNAPSHOT.jar
-
-if sbt has trouble finding jars, add them to the CLASSPATH, andOr check
-./project/build/bioscala.scala, where the automatic load path is set.
-
-= Trouble shooting =
-
-- Error: value biojava is not a member of package org
-
-biojava.jar is not in the class path. Run 'sbt update'
-
-- Error: value should is not a member of java.lang.String
-[error] "SegmentizeGappedSequence" should "handle two"
-
-scalatest-ver.jar is not in the class path. Run 'sbt update'
-
-- value ruby is not a member of package bio
-
-jruby-ver.jar is not in the class path. Add the jar to the class path.
-
