+++ project/build.properties
@@ -1,8 +1,1 @@
-#Project properties
-#Thu May 06 23:18:48 CEST 2010
-project.organization=bioscala
-project.name=bioscala
-sbt.version=0.7.3
-project.version=0.11
-build.scala.versions=2.7.7
-project.initialize=false
+sbt.version=0.13.7-RC2
