+++ projectFilesBackup/.idea_modules/bioscala.iml
@@ -1,0 +1,44 @@
+<module type="JAVA_MODULE" version="4">
+  <component name="FacetManager">
+    <facet type="scala" name="Scala">
+      <configuration>
+        <option name="compilerLibraryLevel" value="Project"/>
+        <option name="compilerLibraryName" value="SBT: scala:2.11.2"/>
+        <option name="languageLevel" value="Scala 2.11"/>
+        <option name="fsc" value="true"/>
+        <option name="compilerOptions" value=""/>
+      </configuration>
+    </facet>
+  </component>
+  <component name="NewModuleRootManager" inherit-compiler-output="false">
+    <output url="file://$MODULE_DIR$/../target/scala-2.11/classes"/>
+    <output-test url="file://$MODULE_DIR$/../target/scala-2.11/test-classes"/>
+    <exclude-output/>
+    <content url="file://$MODULE_DIR$/../">
+      <sourceFolder url="file://$MODULE_DIR$/../src/main/scala" isTestSource="false"/>
+      <sourceFolder url="file://$MODULE_DIR$/../src/main/java" isTestSource="false"/>
+      <sourceFolder url="file://$MODULE_DIR$/../src/main/resources" isTestSource="false"/>
+      <sourceFolder url="file://$MODULE_DIR$/../src/test/scala" isTestSource="true"/>
+      <sourceFolder url="file://$MODULE_DIR$/../src/test/java" isTestSource="true"/>
+      <sourceFolder url="file://$MODULE_DIR$/../src/test/resources" isTestSource="true"/>
+      <excludeFolder url="file://$MODULE_DIR$/../target"/>
+    </content>
+    <orderEntry type="inheritedJdk"/>
+    <orderEntry type="sourceFolder" forTests="false"/>
+    <orderEntry type="library" name="SBT: org.scala-lang:scala-library:2.11.2" level="project"/>
+    <orderEntry type="library" name="SBT: org.biojava:biojava3-core:3.0" level="project"/>
+    <orderEntry type="library" name="SBT: org.biojava:sequencing:1.9.1" level="project"/>
+    <orderEntry type="library" name="SBT: com.google.guava:guava:17.0" level="project"/>
+    <orderEntry type="library" name="SBT: org.biojava:core:1.9.1" level="project"/>
+    <orderEntry type="library" name="SBT: commons-dbcp:commons-dbcp:1.2.2" level="project"/>
+    <orderEntry type="library" name="SBT: commons-pool:commons-pool:1.3" level="project"/>
+    <orderEntry type="library" name="SBT: org.biojava:bytecode:1.9.1" level="project"/>
+    <orderEntry type="library" name="SBT: org.biojava:blast:1.9.1" level="project"/>
+    <orderEntry type="library" name="SBT: com.nativelibs4java:scalaxy-streams_2.11:0.3.2:provided" scope="PROVIDED" level="project"/>
+    <orderEntry type="library" name="SBT: org.scala-lang:scala-compiler:2.11.2:provided" scope="PROVIDED" level="project"/>
+    <orderEntry type="library" name="SBT: org.scala-lang:scala-reflect:2.11.2:provided" scope="PROVIDED" level="project"/>
+    <orderEntry type="library" name="SBT: org.scala-lang.modules:scala-xml_2.11:1.0.2:provided" scope="PROVIDED" level="project"/>
+    <orderEntry type="library" name="SBT: org.scala-lang.modules:scala-parser-combinators_2.11:1.0.2:provided" scope="PROVIDED" level="project"/>
+    <orderEntry type="library" name="SBT: org.scalatest:scalatest_2.11:2.2.1:test" scope="TEST" level="project"/>
+  </component>
+</module>
