+++ src/main/scala/bio/alignment/alignment.scala
@@ -1,3 +1,5 @@
+package bio
+
 /**
  * An Alignment has an (immutable) list of GappedSequences. A Sequence can
  * consist of nucleotides, amino acids and/or gaps. Each of these can have
@@ -5,10 +7,7 @@
  * attributes.
  */
 
-import bio._
-import bio.attribute._
 
-package bio {
 
   class Alignment[T](sequencelist: List[List[T]]) {
     val list = sequencelist
@@ -51,4 +50,48 @@ package bio {
       val gap = Gap
     } 
   } // DNA
-}
+
+/**
+ * Trait contains methods for modifying sparse alignments
+ */
+abstract trait SparseAlignment[T] {
+  type GapType  // Abstract type
+  val gap : GapType
+
+  type Alignment = List[List[T]]  // where T is a bio.Symbol
+
+  def transpose(m: Alignment) : Alignment = {
+    if (m == Nil)
+      Nil
+    else
+    if(m.head.isEmpty) Nil else m.map(_.head) :: transpose(m.map(_.tail))
+  }
+
+  /**
+   * Remove sparse rows, i.e. rows that contain less than min_symbols
+   */
+  def removeSparseRows(in: Alignment, min_symbols: Int) : (Alignment, List[Int]) = {
+    // add an index to the list
+    val indexed = in.zipWithIndex
+    // filter on symbols
+    val filtered = indexed.filter { case (list, index) =>
+      val numgaps = list.count{ i => i == gap }
+      val numsymbols = list.length - numgaps
+      numsymbols >= min_symbols
+    }
+    // and put it all back together again
+    val (ret , idx ) = filtered.unzip
+    val origidx = (0 to in.length-1).toList
+    (ret, origidx.diff(idx)   )
+  }
+
+  /**
+   * Remove columns, i.e. columns that contain less than min_symbols
+   */
+  def removeSparseColumns(in: Alignment, min_symbols: Int) : (Alignment, List[Int]) = {
+    // transpose matrix for optimal list performance
+    val (tm, log) = removeSparseRows(transpose(in),min_symbols)
+    (transpose(tm), log)
+  }
+
+} // SparseAlignment
